Move calculate_status_damage to calculate
